// frontend/src/app/core/api/generated/models/work-order-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface WorkOrderResponse {
  assetId?: string;
  createdAt?: string;
  description?: string;
  id?: string;
  incidentId?: string;
  location?: string;
  priority?: 'P1' | 'P2' | 'P3' | 'P4';
  status?: 'Draft' | 'Ready' | 'In Progress' | 'Done';
  title?: string;
}